+++ build.rs
@@ -1,0 +1,1 @@
+fn main() {println!("cargo:rustc-env=TARGET={}", std::env::var("TARGET").unwrap());}