+++ src/defaults.rs
@@ -7,3 +7,8 @@ pub const DEF_CFG_PATH: &'static str = "/etc/yggdrasil.conf";
 pub const DEF_SOCKET_ADDR: &'static str = "localhost:9001";
 #[cfg(not(target_os = "windows"))]
 pub const DEF_SOCKET_ADDR: &'static str = "/var/run/yggdrasil.sock";
+
+#[cfg(target_os = "windows")]
+pub const EXE_NAME: &'static str = "peers_updater.exe";
+#[cfg(not(target_os = "windows"))]
+pub const EXE_NAME: &'static str = "peers_updater";